Connect to an SQLite datbase

Together, SQLite() and dbConnect() allow you to connect to a SQLite database file. See sqlite-query for how to issue queries and receive results.

SQLite(...)

# S4 method for SQLiteDriver
dbConnect(drv, dbname = "", ...,
  loadable.extensions = TRUE, cache_size = NULL, synchronous = "off",
  flags = SQLITE_RWC, vfs = NULL)

# S4 method for SQLiteConnection
dbConnect(drv, ...)

# S4 method for SQLiteConnection
dbDisconnect(conn, ...)

Arguments

...

In previous versions, SQLite() took arguments. These have now all been moved to dbConnect(), and any arguments here will be ignored with a warning.

drv, conn

An objected generated by SQLite(), or an existing SQLiteConnection. If an connection, the connection will be cloned.

dbname

The path to the database file. SQLite keeps each database instance in one single file. The name of the database is the file name, thus database names should be legal file names in the running platform. There are two exceptions:

"" will create a temporary on-disk database. The file will be deleted when the connection is closed.

":memory:" or "file::memory:" will create a temporary in-memory database.

loadable.extensions

When TRUE (default) SQLite3 loadable extensions are enabled. Setting this value to FALSE prevents extensions from being loaded.

cache_size

Advanced option. A positive integer to change the maximum number of disk pages that SQLite holds in memory (SQLite's default is 2000 pages). See http://www.sqlite.org/pragma.html#pragma_cache_size for details.

synchronous

Advanced options. Possible values for synchronous are "off" (the default), "normal", or "full". Users have reported significant speed ups using sychronous = "off", and the SQLite documentation itself implies considerable improved performance at the very modest risk of database corruption in the unlikely case of the operating system (not the R application) crashing. See http://www.sqlite.org/pragma.html#pragma_synchronous for details.

flags

SQLITE_RWC: open the database in read/write mode and create the database file if it does not already exist; SQLITE_RW: open the database in read/write mode. Raise an error if the file does not already exist; SQLITE_RO: open the database in read only mode. Raise an error if the file does not already exist

vfs

Select the SQLite3 OS interface. See http://www.sqlite.org/vfs.html for details. Allowed values are "unix-posix", "unix-unix-afp", "unix-unix-flock", "unix-dotfile", and "unix-none".

Details

Connections are automatically cleaned-up after they're deleted and reclaimed by the GC. You can use dbDisconnect() to terminate the connection early, but it will not actually close until all open result sets have been closed (and you'll get a warning message to this effect).

Examples

library(DBI)
# Initialize a temporary in memory database and copy a data.frame into it
con <- dbConnect(RSQLite::SQLite(), ":memory:")
data(USArrests)
dbWriteTable(con, "USArrests", USArrests)

#> [1] TRUE

dbListTables(con)

#> [1] "USArrests"

# Fetch all query results into a data frame:
dbGetQuery(con, "SELECT * FROM USArrests")

#>                Murder Assault UrbanPop Rape
#> Alabama          13.2     236       58 21.2
#> Alaska           10.0     263       48 44.5
#> Arizona           8.1     294       80 31.0
#> Arkansas          8.8     190       50 19.5
#> California        9.0     276       91 40.6
#> Colorado          7.9     204       78 38.7
#> Connecticut       3.3     110       77 11.1
#> Delaware          5.9     238       72 15.8
#> Florida          15.4     335       80 31.9
#> Georgia          17.4     211       60 25.8
#> Hawaii            5.3      46       83 20.2
#> Idaho             2.6     120       54 14.2
#> Illinois         10.4     249       83 24.0
#> Indiana           7.2     113       65 21.0
#> Iowa              2.2      56       57 11.3
#> Kansas            6.0     115       66 18.0
#> Kentucky          9.7     109       52 16.3
#> Louisiana        15.4     249       66 22.2
#> Maine             2.1      83       51  7.8
#> Maryland         11.3     300       67 27.8
#> Massachusetts     4.4     149       85 16.3
#> Michigan         12.1     255       74 35.1
#> Minnesota         2.7      72       66 14.9
#> Mississippi      16.1     259       44 17.1
#> Missouri          9.0     178       70 28.2
#>  [ reached getOption("max.print") -- omitted 25 rows ]

# Or do it in batches
rs <- dbSendQuery(con, "SELECT * FROM USArrests")
d1 <- dbFetch(rs, n = 10)      # extract data in chunks of 10 rows
dbHasCompleted(rs)

#> [1] FALSE

d2 <- dbFetch(rs, n = -1)      # extract all remaining data
dbHasCompleted(rs)

#> [1] TRUE

dbClearResult(rs)

#> [1] TRUE

# clean up
dbDisconnect(con)

#> [1] TRUE