dbIsValid {RSQLite}	R Documentation

Check whether an SQLite object is valid or not.

Description

Support function that verifies that the holding a reference to a foreign object is still valid for communicating with the RDBMS

Usage

## S4 method for signature 'SQLiteDriver'
dbIsValid(dbObj)

## S4 method for signature 'SQLiteConnection'
dbIsValid(dbObj)

## S4 method for signature 'SQLiteResult'
dbIsValid(dbObj)

isIdCurrent(obj)

Arguments

dbObj,obj
A driver, connection or result.

Value

A logical scalar.

Examples

dbIsValid(SQLite())

con <- dbConnect(SQLite())
dbIsValid(con)

dbDisconnect(con)
dbIsValid(con)

[Package RSQLite version 1.0.0 Index]