Read a database table

Returns the contents of a database table given by name as a data frame.

# S4 method for SQLiteConnection,character
dbReadTable(conn, name, ...,
  row.names = NA, check.names = TRUE, select.cols = "*")

Arguments

conn

a SQLiteConnection object, produced by DBI::dbConnect()

name

a character string specifying a table name. SQLite table names are not case sensitive, e.g., table names ABC and abc are considered equal.

...

Needed for compatibility with generic. Otherwise ignored.

row.names

Either TRUE, FALSE, NA or a string. If TRUE, always translate row names to a column called "row_names". If FALSE, never translate row names. If NA, translate rownames only if they're a character vector. A string is equivalent to TRUE, but allows you to override the default name. For backward compatibility, NULL is equivalent to FALSE.

check.names

If TRUE, the default, column names will be converted to valid R identifiers.

select.cols

A SQL expression (in the form of a character vector of length 1) giving the columns to select. E.g. "*" selects all columns, "x, y, z" selects three columns named as listed.

Value

A data frame.

Details

Note that the data frame returned by dbReadTable() only has primitive data, e.g., it does not coerce character data to factors.

See also

The corresponding generic function DBI::dbReadTable().

Examples

library(DBI)
db <- RSQLite::datasetsDb()
dbReadTable(db, "mtcars")

#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#>  [ reached getOption("max.print") -- omitted 23 rows ]

dbReadTable(db, "mtcars", row.names = FALSE)

#>              row_names  mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1            Mazda RX4 21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> 2        Mazda RX4 Wag 21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> 3           Datsun 710 22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> 4       Hornet 4 Drive 21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 5    Hornet Sportabout 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 6              Valiant 18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> 7           Duster 360 14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> 8            Merc 240D 24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#>  [ reached getOption("max.print") -- omitted 24 rows ]

dbReadTable(db, "mtcars", select.cols = "cyl, gear")

#>    cyl gear
#> 1    6    4
#> 2    6    4
#> 3    4    4
#> 4    6    3
#> 5    8    3
#> 6    6    3
#> 7    8    3
#> 8    4    4
#> 9    4    4
#> 10   6    4
#> 11   6    4
#> 12   8    3
#> 13   8    3
#> 14   8    3
#> 15   8    3
#> 16   8    3
#> 17   8    3
#> 18   4    4
#> 19   4    4
#> 20   4    4
#> 21   4    3
#> 22   8    3
#> 23   8    3
#> 24   8    3
#> 25   8    3
#> 26   4    4
#> 27   4    5
#> 28   4    5
#> 29   8    5
#> 30   6    5
#> 31   8    5
#> 32   4    4

dbReadTable(db, "mtcars", select.cols = "row_names, cyl, gear")

#>                     cyl gear
#> Mazda RX4             6    4
#> Mazda RX4 Wag         6    4
#> Datsun 710            4    4
#> Hornet 4 Drive        6    3
#> Hornet Sportabout     8    3
#> Valiant               6    3
#> Duster 360            8    3
#> Merc 240D             4    4
#> Merc 230              4    4
#> Merc 280              6    4
#> Merc 280C             6    4
#> Merc 450SE            8    3
#> Merc 450SL            8    3
#> Merc 450SLC           8    3
#> Cadillac Fleetwood    8    3
#> Lincoln Continental   8    3
#> Chrysler Imperial     8    3
#> Fiat 128              4    4
#> Honda Civic           4    4
#> Toyota Corolla        4    4
#> Toyota Corona         4    3
#> Dodge Challenger      8    3
#> AMC Javelin           8    3
#> Camaro Z28            8    3
#> Pontiac Firebird      8    3
#> Fiat X1-9             4    4
#> Porsche 914-2         4    5
#> Lotus Europa          4    5
#> Ford Pantera L        8    5
#> Ferrari Dino          6    5
#> Maserati Bora         8    5
#> Volvo 142E            4    4

dbDisconnect(db)

#> [1] TRUE