initRegExp {RSQLite}	R Documentation

Add regular expression operator

Description

This loads a regular-expression matcher for posix extended regular expressions, as available through the SQLite source code repository (https://sqlite.org/src/raw?filename=ext/misc/regexp.c).

Usage

initRegExp(db)

Arguments

db
A SQLiteConnection object to add the regular expression operator into the connection.

Details

SQLite will then implement the "A regexp B" operator, where A is the string to be matched and B is the regular expression.

Note this only affects the specified connection.

Examples

library(DBI)
db <- RSQLite::datasetsDb()
RSQLite::initRegExp(db)

dbGetQuery(db, "SELECT * FROM mtcars WHERE carb REGEXP '[12]'")
dbDisconnect(db)

[Package RSQLite version 2.1.2.9000 Index]