Result information

For a result object, returns information about the SQL statement used, the available columns and number of already fetched rows for a query, the number of affected rows for a statement, and the completion status.

# S4 method for SQLiteResult
dbColumnInfo(res, ...)

# S4 method for SQLiteResult
dbGetRowsAffected(res, ...)

# S4 method for SQLiteResult
dbGetRowCount(res, ...)

# S4 method for SQLiteResult
dbHasCompleted(res, ...)

# S4 method for SQLiteResult
dbGetStatement(res, ...)

Arguments

res
An object of class SQLiteResult

...
Ignored. Needed for compatibility with generic

See also

The corresponding generic functions DBI::dbColumnInfo(), DBI::dbGetRowsAffected(), DBI::dbGetRowCount(), DBI::dbHasCompleted(), and DBI::dbGetStatement().

Examples

library(DBI)
db <- RSQLite::datasetsDb()
rs <- dbSendQuery(db, "SELECT * FROM USArrests WHERE UrbanPop >= 80")
dbGetStatement(rs)

#> [1] "SELECT * FROM USArrests WHERE UrbanPop >= 80"

dbColumnInfo(rs)

#>        name        type
#> 1 row_names     promise
#> 2    Murder environment
#> 3   Assault    pairlist
#> 4  UrbanPop    pairlist
#> 5      Rape environment

dbHasCompleted(rs)

#> [1] FALSE

dbGetRowCount(rs)

#> [1] 0

dbFetch(rs, n = 2)

#>            Murder Assault UrbanPop Rape
#> Arizona       8.1     294       80 31.0
#> California    9.0     276       91 40.6

dbHasCompleted(rs)

#> [1] FALSE

dbGetRowCount(rs)

#> [1] 2

invisible(dbFetch(rs))
dbHasCompleted(rs)

#> [1] TRUE

dbGetRowCount(rs)

#> [1] 12

dbClearResult(rs)

dbDisconnect(db)

con <- dbConnect(RSQLite::SQLite(), ":memory:")
dbExecute(con, "CREATE TABLE test (a INTEGER)")

#> [1] 0

rs <- dbSendStatement(con, "INSERT INTO test VALUES (:a)", list(a = 1:3))
dbGetRowsAffected(rs)

#> [1] 3

dbClearResult(rs)
dbDisconnect(con)