Execute a SQL statement on a database connection

To retrieve results a chunk at a time, use dbSendQuery(), dbFetch(), then dbClearResult(). Alternatively, if you want all the results (and they'll fit in memory) use dbGetQuery() which sends, fetches and clears for you. To run the same prepared query with multiple inputs, use dbBind().

# S4 method for SQLiteConnection,character
dbSendQuery(conn, statement,
  params = NULL, ...)

# S4 method for SQLiteResult
dbBind(res, params, ...)

# S4 method for SQLiteResult
dbFetch(res, n = -1, ..., row.names = NA)

# S4 method for SQLiteResult
dbClearResult(res, ...)

Arguments

conn

an SQLiteConnection object.

statement

a character vector of length one specifying the SQL statement that should be executed. Only a single SQL statment should be provided.

params

A named list of query parameters to be substituted into a parameterised query.

...

Unused. Needed for compatibility with generic.

res

an SQLiteResult object.

n

maximum number of records to retrieve per fetch. Use -1 to retrieve all pending records; use 0 for to fetch the default number of rows as defined in SQLite()

row.names

Either TRUE, FALSE, NA or a string. If TRUE, always translate row names to a column called "row_names". If FALSE, never translate row names. If NA, translate rownames only if they're a character vector. A string is equivalent to TRUE, but allows you to override the default name. For backward compatibility, NULL is equivalent to FALSE.

Details

See sqlite-meta for how to extract other metadata from the result set.

Examples

library(DBI)
db <- RSQLite::datasetsDb()

# Run query to get results as dataframe
dbGetQuery(db, "SELECT * FROM USArrests LIMIT 3")

#>         Murder Assault UrbanPop Rape
#> Alabama   13.2     236       58 21.2
#> Alaska    10.0     263       48 44.5
#> Arizona    8.1     294       80 31.0

# Send query to pull requests in batches
res <- dbSendQuery(db, "SELECT * FROM USArrests")
dbFetch(res, n = 2)

#>         Murder Assault UrbanPop Rape
#> Alabama   13.2     236       58 21.2
#> Alaska    10.0     263       48 44.5

dbFetch(res, n = 2)

#>          Murder Assault UrbanPop Rape
#> Arizona     8.1     294       80 31.0
#> Arkansas    8.8     190       50 19.5

dbHasCompleted(res)

#> [1] FALSE

dbClearResult(res)

#> [1] TRUE

# Parameterised queries are safest when you accept user input
dbGetQuery(db, "SELECT * FROM USArrests WHERE Murder < ?", list(3))

#>               Murder Assault UrbanPop Rape
#> Idaho            2.6     120       54 14.2
#> Iowa             2.2      56       57 11.3
#> Maine            2.1      83       51  7.8
#> Minnesota        2.7      72       66 14.9
#> New Hampshire    2.1      57       56  9.5
#> North Dakota     0.8      45       44  7.3
#> Vermont          2.2      48       32 11.2
#> Wisconsin        2.6      53       66 10.8

# Or create and then bind
rs <- dbSendQuery(db, "SELECT * FROM USArrests WHERE Murder < ?")
dbBind(rs, list(3))
dbFetch(rs)

#>               Murder Assault UrbanPop Rape
#> Idaho            2.6     120       54 14.2
#> Iowa             2.2      56       57 11.3
#> Maine            2.1      83       51  7.8
#> Minnesota        2.7      72       66 14.9
#> New Hampshire    2.1      57       56  9.5
#> North Dakota     0.8      45       44  7.3
#> Vermont          2.2      48       32 11.2
#> Wisconsin        2.6      53       66 10.8

# Named parameters are a little more convenient
rs <- dbSendQuery(db, "SELECT * FROM USArrests WHERE Murder < :x")

#> Warning: Closing open result set, pending rows

dbBind(rs, list(x = 3))
dbFetch(rs)

#>               Murder Assault UrbanPop Rape
#> Idaho            2.6     120       54 14.2
#> Iowa             2.2      56       57 11.3
#> Maine            2.1      83       51  7.8
#> Minnesota        2.7      72       66 14.9
#> New Hampshire    2.1      57       56  9.5
#> North Dakota     0.8      45       44  7.3
#> Vermont          2.2      48       32 11.2
#> Wisconsin        2.6      53       66 10.8

dbDisconnect(db)

#> Warning: There are 1 result in use. The connection will be released when they are closed

#> [1] TRUE