sqliteQuickColumn {RSQLite}	R Documentation

Return an entire column from a SQLite database

Description

Return an entire column from a table in a SQLite database as an R vector of the appropriate type. This function is experimental and subject to change.

Usage

sqliteQuickColumn(con, table, column)

Arguments

con
a SQLiteConnection object as produced by sqliteNewConnection.

table
a string specifying the name of the table

column
a string specifying the name of the column in the specified table to retrieve.

Details

This function relies upon the SQLite internal ROWID column to determine the number of rows in the table. This may not work depending on the table schema definition and pattern of update.

Value

an R vector of the appropriate type (based on the type of the column in the database).

Author(s)

Seth Falcon

[Package RSQLite version 1.0.0 Index]