dbReadTable,SQLiteConnection,character-method {RSQLite}	R Documentation

Convenience functions for importing/exporting DBMS tables

Description

These functions mimic their R/S-Plus counterpart get, assign, exists, remove, and objects, except that they generate code that gets remotely executed in a database engine.

Usage

## S4 method for signature 'SQLiteConnection,character'
dbReadTable(conn, name, row.names,
  check.names = TRUE, select.cols = "*")

Arguments

conn
a SQLiteConnection object, produced by dbConnect

name
a character string specifying a table name. SQLite table names are not case sensitive, e.g., table names ABC and abc are considered equal.

row.names
A string or an index specifying the column in the DBMS table to use as row.names in the output data.frame. Defaults to using the row_names column if present. Set to NULL to never use row names.

check.names
If TRUE, the default, column names will be converted to valid R identifiers.

select.cols
A SQL statement (in the form of a character vector of length 1) giving the columns to select. E.g. "*" selects all columns, "x,y,z" selects three columns named as listed.

Value

A data.frame in the case of dbReadTable; otherwise a logical indicating whether the operation was successful.

Note

Note that the data.frame returned by dbReadTable only has primitive data, e.g., it does not coerce character data to factors.

Examples

con <- dbConnect(SQLite())
dbWriteTable(con, "mtcars", mtcars)
dbReadTable(con, "mtcars")

# Supress row names
dbReadTable(con, "mtcars", row.names = FALSE)

dbDisconnect(con)

[Package RSQLite version 1.0.0 Index]