initExtension {RSQLite}	R Documentation

Add useful extension functions.

Description

These extension functions are written by Liam Healy and made available through the SQLite website (http://www.sqlite.org/contrib).

Usage

initExtension(db)

Arguments

db
A database to load these extensions.

Available extension functions

Math functions

acos, acosh, asin, asinh, atan, atan2, atanh, atn2, ceil, cos, cosh, cot, coth, degrees, difference, exp, floor, log, log10, pi, power, radians, sign, sin, sinh, sqrt, square, tan, tanh

String functions

charindex, leftstr, ltrim, padc, padl, padr, proper, replace, replicate, reverse, rightstr, rtrim, strfilter, trim

Aggregate functions

stdev, variance, mode, median, lower_quartile, upper_quartile

Examples

db <- dbConnect(SQLite())
initExtension(db)

dbWriteTable(db, "mtcars", mtcars)
dbGetQuery(db, "SELECT stdev(mpg) FROM mtcars")
sd(mtcars$mpg)

[Package RSQLite version 1.0.0 Index]