sqliteCopyDatabase {RSQLite}	R Documentation

Copy a SQLite database

Description

Copies a database connection to a file or to another database connection. It can be used to save an in-memory database (created using dbname = ":memory:" or dbname = "file::memory:") to a file or to create an in-memory database a copy of another database.

Usage

sqliteCopyDatabase(from, to)

Arguments

from
A SQLiteConnection object. The main database in from will be copied to to.

to
A SQLiteConnection object pointing to an empty database.

Author(s)

Seth Falcon

References

http://www.sqlite.org/backup.html

Examples

library(DBI)
# Copy the built in databaseDb() to an in-memory database
con <- dbConnect(RSQLite::SQLite(), ":memory:")
dbListTables(con)

db <- RSQLite::datasetsDb()
RSQLite::sqliteCopyDatabase(db, con)
dbDisconnect(db)
dbListTables(con)

dbDisconnect(con)

[Package RSQLite version 2.1.2.9000 Index]